chore: refactor progress
